chore: move all helpers in a dedicated NS
