+++ src/Console/Application.php
@@ -20,9 +20,9 @@ use Castor\ExpressionLanguage;
 use Castor\Fingerprint\FingerprintHelper;
 use Castor\FunctionFinder;
 use Castor\GlobalHelper;
+use Castor\Helper\PlatformHelper;
+use Castor\Helper\WaitForHelper;
 use Castor\Import\Importer;
-use Castor\PlatformHelper;
-use Castor\WaitForHelper;
 use Monolog\Logger;
 use Psr\Cache\CacheItemPoolInterface;
 use Psr\Log\LogLevel;
