+++ src/Console/ApplicationFactory.php
@@ -10,6 +10,9 @@ use Castor\EventDispatcher;
 use Castor\ExpressionLanguage;
 use Castor\Fingerprint\FingerprintHelper;
 use Castor\FunctionFinder;
+use Castor\Helper\PathHelper;
+use Castor\Helper\PlatformHelper;
+use Castor\Helper\WaitForHelper;
 use Castor\Import\Importer;
 use Castor\Import\Listener\RemoteImportListener;
 use Castor\Import\Remote\Composer;
@@ -17,10 +20,7 @@ use Castor\Import\Remote\PackageImporter;
 use Castor\Listener\GenerateStubsListener;
 use Castor\Listener\UpdateCastorListener;
 use Castor\Monolog\Processor\ProcessProcessor;
-use Castor\PathHelper;
-use Castor\PlatformHelper;
 use Castor\Stub\StubsGenerator;
-use Castor\WaitForHelper;
 use Monolog\Logger;
 use Symfony\Component\Cache\Adapter\FilesystemAdapter;
 use Symfony\Component\Console\Application as SymfonyApplication;
