+++ src/FunctionFinder.php
@@ -13,6 +13,7 @@ use Castor\Descriptor\ListenerDescriptor;
 use Castor\Descriptor\SymfonyTaskDescriptor;
 use Castor\Descriptor\TaskDescriptor;
 use Castor\Exception\FunctionConfigurationException;
+use Castor\Helper\SluggerHelper;
 use Symfony\Component\Console\Attribute\AsCommand;
 use Symfony\Component\Finder\Finder;
 use Symfony\Component\Process\Process;
