+++ src/HasherHelper.php
@@ -2,161 +2,13 @@
 
 namespace Castor;
 
-use Castor\Console\Application;
-use Castor\Fingerprint\FileHashStrategy;
-use Psr\Log\LoggerInterface;
-use Psr\Log\NullLogger;
-use Symfony\Component\Finder\Finder;
+use Castor\Helper\HasherHelper as BaseHasherHelper;
 
-class HasherHelper
+trigger_deprecation('castor', '0.15', 'The "%s" class is deprecated, use "%s" instead.', HasherHelper::class, BaseHasherHelper::class);
+
+/**
+ * @deprecated since castor/castor 0.15, use Castor\Helper\HasherHelper instead.
+ */
+class HasherHelper extends BaseHasherHelper
 {
-    private readonly \HashContext $hashContext;
-
-    /**
-     * @see https://www.php.net/manual/en/function.hash-algos.php
-     */
-    public function __construct(
-        private readonly Application $application,
-        private readonly LoggerInterface $logger = new NullLogger(),
-        string $algo = 'xxh128',
-    ) {
-        $this->hashContext = hash_init($algo);
-    }
-
-    public function write(string $value): self
-    {
-        $this->logger->debug('Hashing value "{value}".', ['value' => $value]);
-
-        hash_update($this->hashContext, $value);
-
-        return $this;
-    }
-
-    public function writeFile(string $path, FileHashStrategy $strategy = FileHashStrategy::MTimes): self
-    {
-        if (!str_starts_with($path, '/')) {
-            $path = getcwd() . '/' . $path;
-        }
-
-        if (!is_file($path)) {
-            throw new \InvalidArgumentException(sprintf('The path "%s" is not a file.', $path));
-        }
-
-        if (!is_readable($path)) {
-            throw new \InvalidArgumentException(sprintf('The file "%s" is not readable.', $path));
-        }
-
-        $this->logger->debug('Hashing file "{path}" with strategy "{strategy}".', [
-            'path' => $path,
-            'strategy' => $strategy->name,
-        ]);
-
-        switch ($strategy) {
-            case FileHashStrategy::Content:
-                hash_update_file($this->hashContext, $path);
-
-                break;
-            case FileHashStrategy::MTimes:
-                hash_update($this->hashContext, sprintf('%s:%s', $path, filemtime($path)));
-
-                break;
-        }
-
-        return $this;
-    }
-
-    public function writeWithFinder(Finder $finder, FileHashStrategy $strategy = FileHashStrategy::MTimes): self
-    {
-        $this->logger->debug('Hashing files with Finder with strategy "{strategy}".', [
-            'strategy' => $strategy->name,
-        ]);
-
-        foreach ($finder as $file) {
-            switch ($strategy) {
-                case FileHashStrategy::Content:
-                    hash_update_file($this->hashContext, $file->getPathname());
-
-                    break;
-                case FileHashStrategy::MTimes:
-                    hash_update($this->hashContext, "{$file->getPathname()}:{$file->getMTime()}");
-
-                    break;
-            }
-        }
-
-        return $this;
-    }
-
-    public function writeGlob(string $pattern, FileHashStrategy $strategy = FileHashStrategy::MTimes): self
-    {
-        $this->logger->debug('Hashing files {pattern} with strategy "{strategy}".', [
-            'pattern' => $pattern,
-            'strategy' => $strategy->name,
-        ]);
-
-        $files = glob($pattern);
-
-        if (false === $files) {
-            throw new \InvalidArgumentException(sprintf('The pattern "%s" is invalid.', $pattern));
-        }
-
-        foreach ($files as $file) {
-            switch ($strategy) {
-                case FileHashStrategy::Content:
-                    hash_update_file($this->hashContext, $file);
-
-                    break;
-                case FileHashStrategy::MTimes:
-                    $modifiedTime = filemtime($file);
-                    hash_update($this->hashContext, "{$file}:{$modifiedTime}");
-
-                    break;
-            }
-        }
-
-        return $this;
-    }
-
-    public function writeTaskName(): self
-    {
-        $taskName = $this->application->getCommand()->getName() ?? 'n/a';
-
-        $this->logger->debug('Hashing task name "{name}".', [
-            'name' => $taskName,
-        ]);
-
-        hash_update($this->hashContext, $taskName);
-
-        return $this;
-    }
-
-    public function writeTaskArgs(string ...$args): self
-    {
-        $this->logger->debug('Hashing task args "{args}".', [
-            'args' => implode(', ', $args),
-        ]);
-
-        $input = $this->application->getInput();
-
-        foreach ($args as $arg) {
-            if ($input->hasArgument($arg)) {
-                $this->write($input->getArgument($arg));
-            }
-        }
-
-        return $this;
-    }
-
-    public function writeTask(string ...$args): self
-    {
-        $this->writeTaskName();
-        $this->writeTaskArgs(...$args);
-
-        return $this;
-    }
-
-    public function finish(): string
-    {
-        return hash_final($this->hashContext);
-    }
 }
