+++ src/Import/Remote/Composer.php
@@ -5,8 +5,8 @@ namespace Castor\Import\Remote;
 use Castor\Console\Application;
 use Castor\Fingerprint\FingerprintHelper;
 use Castor\GlobalHelper;
+use Castor\Helper\PathHelper;
 use Castor\Import\Exception\ComposerError;
-use Castor\PathHelper;
 use Psr\Log\LoggerInterface;
 use Symfony\Component\Console\Helper\ProgressIndicator;
 use Symfony\Component\Filesystem\Filesystem;
